% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_1D.R
\name{fft_1D}
\alias{fft_1D}
\title{Fourier Transform in one dimension}
\usage{
fft_1D(X, Y)
}
\arguments{
\item{X}{The X column of the 1-D dataset}

\item{Y}{The Y column of the 1-D dataset}
}
\value{
The returned value is always a Three column dataframe .\cr
\itemize{
\item The first column contain \eqn{\large \omega_k}
\item The second column contains the amplitude of the complex \eqn{\large X_k}, \eqn{\large ||X_k||}
\item The third column contains the phase vector of the complex \eqn{\large X_k}
}
}
\description{
The function takes in vectors \eqn{\bar{X}} and \eqn{\bar{Y}} and returns the Fourier Transfrom of the data
}
\details{
Solves for the DFT, \eqn{\large X_k}, using Fortran's Fast FFTW3 backend \link{FFTW3}\url{https://www.fftw.org/}. \cr
\deqn{\displaystyle \large X_k = \sum_{n=0}^{N-1}x_n e^{-2\pi i k n/N}}
Calculates the frequency vector, \eqn{\large \omega_k} using the sampling of the \eqn{\large \bar{X}} variable.\cr
Here, \eqn{ \large \omega_0 \leq \omega_k \leq \omega_s} where \eqn{\large \omega_0} and \eqn{\large \omega_s} are the base and the sampling frequencies.\cr
}
\examples{
#Call the relevant libraries
library(StatsChitran)
library(ImageMetrologyAnalysis)

#build the dataset
X <- seq(-10, 10, 0.1)
Y1 <- gauss.1D(X)
Y2 <- gauss.1D(X, mu = 0, sig = 1.5)
Y3 <- gauss.1D(X, mu = 0, sig = 2.0)
plot(X, Y1, type='l', col = 'black')
lines(X, Y2, col = 'blue')
lines(X, Y3, col = 'red')


}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
