% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_2D.R
\name{fft_2D}
\alias{fft_2D}
\title{Fourier Transform in two dimensions}
\usage{
fft_2D(tens, sampling.del, pl)
fft_2D(tens, sampling.del)
fft_2D(tens)
}
\arguments{
\item{tens}{The tensor describing the lattice for which the DFT is needed}

\item{sampling.del}{The allowed relative sampling error difference along the X and Y vectors w.r.t the \code{tens} argument\cr
\itemize{
\item For reliably displaying the 2D DFT, the sampling frequencies along X and Y need to be the same \cr
\item If the sampling error, \code{sampling.del} \eqn{> (|\Delta X| - |\Delta Y|)/|\Delta X|}, then an error is thrown \cr
\item Here, \eqn{\Delta X = } \code{mean(diff(X))} and \eqn{\Delta Y = } \code{mean(diff(Y))}
\item \code{sampling.del} defaults to \eqn{0.1}
}}

\item{pl}{Plots either the amplitude spectrum, the phase spectrum or nothing.\cr
The accepted inputs are \code{pl = 'amp'}, \code{pl = 'phase'} or \code{pl = 'none'}\cr}
}
\value{
The returned value is always a Two element list.\cr
\itemize{
\item The first element contains the amplitude tensor
\item The second element contains the phase tensor
}
}
\description{
The function takes in a rank-3 tensor of the form \eqn{\large T_{m,n,p = 3}} and returns the Fourier Lattice of the tensor
}
\details{
Solves for the 2D DFT, \eqn{\large X_{k,l}}, using Fortran's Fast FFTW3 backend \url{https://www.fftw.org/}. \cr
\deqn{\displaystyle \Large X_{k,l} = \sum_{m=0}^{M-1}\sum_{n=0}^{N-1}x_{m,n} e^{-2\pi i (k n/N + l m/M)}}
Zero pads the dataset to make \eqn{\large M = N} before DFT computation.\cr
Calculates the frequency vector, \eqn{\large \omega_k} using the sampling of the \eqn{\large \bar{X}} or \eqn{\large \bar{Y}} variables from the \code{tens} argument, sice both are forced to be roughly similar.\cr
Here, \eqn{ \large \omega_0 \leq \omega_k \leq \omega_s} where \eqn{\large \omega_0} and \eqn{\large \omega_s} are the base and the sampling frequencies.\cr
Zero padding with \eqn{\large M = N} forces \eqn{\large \omega_0} to be the same for both fourier frquency axes.\cr
While \eqn{\large \Delta X \sim \Delta Y} forces \eqn{\large \omega_s} to be the same for both fourier frequency axes.\cr
}
\examples{
rm(list = ls())
#Call the relevant libraries
library(ImageMetrologyAnalysis)
library(StatsChitran)
#Build X and Y sequences
X <- seq(-4, 4, by=0.1)
Y <- seq(-5, 5, by = 0.1)
#Build the rectangular lattice
rec.tens <- rect.latt(X, Y, R.latt.x = 0.5, R.latt.y = 0.8, A = 1, sig = 0.15)
#Call the ffts and plot them
f.rec.tens <- fft_2D(tens=rec.tens, pl='amp')
f.rec.tens.amp = f.rec.tens[[1]]
f.rec.tens.amp[,,1] <- log(f.rec.tens.amp[,,1])
plot2D.arr(f.rec.tens.amp)


}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
