% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func2D.R
\name{func2D}
\alias{func2D}
\title{Returns a tensor from a function}
\usage{
func2D(f, X, Y)
}
\arguments{
\item{f}{The function \eqn{f(\bar{X}, \bar{Y})} with caveats\cr
\enumerate{
\item The function \eqn{f(\bar{X}, \bar{Y})} should have two arguments of the form \code{f(X, Y)}
\item The arguments should be named \code{X} and \code{Y}
}}

\item{X}{The vector \eqn{\bar{X}}, or \code{X}, used as an argument in \eqn{f(\bar{X}, \bar{Y})}, or \code{f(X, Y)}}

\item{Y}{The vector \eqn{\bar{Y}}, or \code{Y}, used as an argument in \eqn{f(\bar{X}, \bar{Y})} or \code{f(X, Y)}}
}
\value{
The returned value is always a rank-3 tensor of the form checked by \code{\link{is.tensor.rank3}}
}
\description{
The function takes in a funtion \eqn{f(\bar{X}, \bar{Y})} and two vectors \eqn{\bar{X}}, \eqn{\bar{Y}} and returns a rank-3 tensor of the form \code{\link{is.tensor.rank3}}
}
\details{
Creates a rank-3 tensor out of a 2D funtion \code{f(X, Y)} and vectors \code{X} and \code{Y}
}
\examples{
library(ImageMetrologyAnalysis)
library(StatsChitran)
X <- seq(-10, 10, by = 0.01)
Y <- seq(-8, 8, by = 0.01)
f <- function(X,Y){
  return(sin(3*X+4*Y) + cos(-4*X + 3*Y))
}
tens <- func2D(f, X, Y)
plot2D.arr(tens)


}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
