% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss.1D.R
\name{gauss.1D}
\alias{gauss.1D}
\title{Gaussian in one dimension}
\usage{
gauss.1D(X, A, mu, sig)
gauss.1D(X, mu, sig)
gauss.1D(X, mu)
gauss.1D(X)
}
\arguments{
\item{X}{The numeric vector which contains the \eqn{x} values against which the gaussian vector is needed.}

\item{A}{The term, \eqn{A}, denoting the amplitude of the gaussian.\cr
If not provided, the term defaults to \eqn{A=1/\sigma \sqrt{2\pi}}}

\item{mu}{The term, \eqn{\mu}, denoting the mean of the gaussian.\cr
If not provided, the term defaults to \eqn{\mu = 0}}

\item{sig}{The term, \eqn{\sigma}, denoting the satndard-deviation of the gaussian.\cr
If not provided, the term defaults to \eqn{\sigma = 1}}
}
\value{
The returned value is always a vector
}
\description{
The function takes in a vector \eqn{\bar{X}} against which it returns a gaussian vector depending on other parameters, like Amplitude, mean and standard-deviation.
}
\details{
The function returns a vector \eqn{\bar{Y}} with individual compunents \eqn{y} against the vector \eqn{\bar{X}} with components \eqn{x}\cr
\deqn{\large\displaystyle y = A e^{-(x-\mu)^2 / 2\sigma^2} } \cr
When \eqn{A}, \eqn{\mu} and \eqn{\sigma} are missing, then the returned vector \eqn{\bar{Y}} is the standard normal distribution against \eqn{\bar{X}}\cr
Whenever \eqn{A} is missing, then \eqn{\bar{Y}} is statistically normalized using \eqn{A=1/\sigma \sqrt{2\pi}}\cr
When all the arguments are provided and \eqn{A \neq 1/\sigma \sqrt{2\pi}}, then \eqn{\bar{Y}} remains non-normalized
}
\examples{
#Call the relevant libraries
library(StatsChitran)
library(ImageMetrologyAnalysis)

#build the dataset
X <- seq(-10, 10, 0.1)
Y1 <- gauss.1D(X)
Y2 <- gauss.1D(X, mu = 0, sig = 1.5)
Y3 <- gauss.1D(X, mu = 0, sig = 2.0)
plot(X, Y1, type='l', col = 'black')
plot(X, Y2, type='l', col = 'blue')
plot(X, Y3, type = 'l', col = 'red')


}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
