% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss.2D.nocorr.R
\name{gauss.2D.nocorr}
\alias{gauss.2D.nocorr}
\title{Gaussian in two dimension}
\usage{
gauss.2D.nocorr(X, Y)
gauss.2D.nocorr(X, Y, mu.x, mu.y, sig.x)
gauss.2D.nocorr(X, Y, mu.x, sig.x, sig.y)
gauss.2D.nocorr(X, Y, mu.x, mu.y, sig.x, sig.y)
gauss.2D.nocorr(X, Y, A.x, A.y, mu.x, mu.y, sig.x, sig.y)
}
\arguments{
\item{X, Y}{The numeric vectors \eqn{\large \bar{X}} and \eqn{\large \bar{Y}}}

\item{A.x, A.y}{The terms, \eqn{A_x}, \eqn{A_y}, denoting the amplitudes of the individual gaussian.\cr
If not provided, the terms default to \eqn{A_x=1/\sigma_x \sqrt{2\pi}}, \eqn{A_y=1/\sigma_y \sqrt{2\pi}}}

\item{mu.x, mu.y}{The term, \eqn{\mu_x} and \eqn{\mu_y}, denoting the means of the individual gaussian.\cr
If not provided, the term defaults to \eqn{\mu_x = 0}, \eqn{\mu_y = 0}}

\item{sig.x, sig.y}{The terms, \eqn{\sigma_x} and \eqn{\sigma_y}, denoting the satndard-deviations of the individual gaussian.\cr
If not provided, the terms default to \eqn{\sigma_x = 1}, \eqn{\sigma_y = 1}}
}
\value{
The returned value is always a rank 3 tensor \eqn{\large T_{m,n,p=3}} of the type returned in \link{StatsChitran::img_2_arr} \cr
\itemize{
\item \eqn{\large T_{m,n,1}} is the 2D gaussian mesh
\item \eqn{\large T_{m,n,2}} is the mesh built out of \eqn{\large \bar{X}}
\item \eqn{\large T_{m,n,3}} is the mesh built out of \eqn{\large \bar{Y}}
}
}
\description{
Builds a gaussian mesh out of two \strong{uncorrelated} vectors \eqn{\large \bar{X}} and \eqn{\large \bar{Y}}
}
\details{
The function returns a tensor \eqn{\large T_{m,n,p}} where \eqn{\large T_{m,n,1}} is the gaussian mesh or the result of the gaussian function \eqn{\large f_G(x,y)}\cr
\deqn{\Large\displaystyle f_G(x,y) = A_x A_y e^{-\{(x-\mu_x)^2 / 2{\sigma_x}^2 + (y-\mu_y)^2 / 2{\sigma_y}^2 \}}} \cr
It's important to remember that the gaussians generated here are out of two uncorrelated vectors \eqn{\large \bar{X}} and \eqn{\large \bar{Y}} \cr
When \eqn{\large A_{x,y}}, \eqn{\large \mu_{x,y}} and \eqn{\large \sigma_{x,y}} are missing, then the returned mesh \eqn{\large T_{m,n,1}} is the standard normal distribution in two dimensions \cr
Whenever \eqn{\large A_x} \strong{and} \eqn{\large A_y} are missing, then the mesh \eqn{\large T_{m,n,1}} is statistically normalized using \eqn{\large A_{x,y}=1/\sigma_{x,y} \sqrt{2\pi}}\cr
\eqn{\large T_{m,n,2}} and \eqn{\large T_{m,n,3}} are the meshes from the \eqn{\large \bar{X}} and \eqn{\large \bar{Y}} vectors respectively \cr
}
\examples{
rm(list = ls())
#Call the relevant libraries
library(ImageMetrologyAnalysis)
library(StatsChitran)

#Build the X and Y gaussians
X <- seq(-8, 10, by=0.01)
Y <- seq(-5, 10, by=0.01)
tens <- gauss.2D.nocorr(X, Y, mu.x = 1.0, mu.y = 5.0, sig.x = 2.3) #Gaussian centered at (1,5)
tens1 <- gauss.2D.nocorr(X, Y, sig.y = 1.5) #Gaussian centered at (0,0) default
tens[,,1] <- tens[,,1] + tens1[,,1] #Add the two gaussians

#Plot the array
plot2D.arr(tens)


}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
