% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_2.R
\name{grid_2}
\alias{grid_2}
\title{Returns a 2D X-Y grid/rank-3 tensor}
\usage{
grid_2(X, Y)
}
\arguments{
\item{X}{The vector \eqn{\bar{X}}, or \code{X}.}

\item{Y}{The vector \eqn{\bar{Y}}, or \code{Y}.}
}
\value{
The returned value is always a rank-3 tensor of the form checked by \code{\link{is.tensor.rank3}}
}
\description{
The function takes in two vectors \eqn{\large \bar{X}} and \eqn{\large \bar{Y}} and returns a rank-3 tensor \eqn{\large T_{m,n,p=2}}
}
\details{
Creates a rank-3 tensor \eqn{\large T_{m,n,p=2}} extablishing the X-Y grid
\enumerate{
\item The first slice \eqn{\large T_{m,n,1}} is the X-values of the XY-grid
\item The second slice \eqn{\large T_{m,n,2}} is the Y-values of the XY-grid
}
}
\examples{
library(ImageMetrologyAnalysis)
library(StatsChitran)
X <- seq(1, 5)
Y <- seq(-2, 2)
grid.tens <- grid_2(X, Y)
grid.tens[,,1] #The X-matrix
grid.tens[,,2] #The Y-matrix


}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
