% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.latt.R
\name{hex.latt}
\alias{hex.latt}
\title{Build a rank-3 tensor pertaining to a periodic hexagonal lattice}
\usage{
hex.latt(X, Y, R.latt, A, sig, pl)
hex.latt(X, Y, R.latt, A, sig)
}
\arguments{
\item{X, Y}{The \eqn{\large \bar{X}} and \eqn{\large \bar{Y}} vectors
pertaining to the image spans.\cr
See \code{\link{grid_2}} for usage specifics of the X and Y vectors.}

\item{R.latt}{The lattice constant \eqn{\large R_{\mathrm{latt}}}
defining the periodicity of the hexagonal lattice.}

\item{A, sig}{The amplitude and standard deviation of the Gaussian
placed at each lattice site.}

\item{pl}{When set to \code{TRUE}, plots the rank-3 tensor using
\code{\link[StatsChitran]{plot2D.arr}}.\cr
Defaults to \code{TRUE}.}
}
\value{
The returned value is always a rank-3 tensor.\cr
\itemize{
\item The first slice, \eqn{\large T_{m,n,p=1}}, contains the lattice
Z-values.
\item The second slice, \eqn{\large T_{m,n,p=2}}, contains the X-values
of the X-Y grid.
\item The third slice, \eqn{\large T_{m,n,p=3}}, contains the Y-values
of the X-Y grid.
}
}
\description{
The function takes in vectors \eqn{\bar{X}}, \eqn{\bar{Y}} and the
hexagonal lattice constant \eqn{\large R_{\mathrm{latt}}}, and returns
the rank-3 tensor pertaining to the lattice.\cr
}
\details{
Creates a periodic hexagonal lattice by placing Gaussians at each
lattice site within the \eqn{\large \bar{X}}, \eqn{\large \bar{Y}} span.\cr
The lattice periodicity is defined by the lattice constant
\eqn{\large R_{\mathrm{latt}}}.\cr
}
\examples{
rm(list = ls())
library(ImageMetrologyAnalysis)
library(StatsChitran)

X <- seq(-5, 5, by = 0.01)
Y <- seq(-3, 3, by = 0.01)

hex.tens <- hex.latt(
  X,
  Y,
  R.latt = 0.5,
  A = 1,
  sig = 0.15
)

}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
