% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.function.R
\name{integrate.function}
\alias{integrate.function}
\title{Integrate an entire dataset to get another dataset}
\usage{
integrate.function(X, Y, y0)
}
\arguments{
\item{X}{The numeric vector \eqn{\large \bar{X}}, such that \eqn{\large x[k] \in \bar{X}}}

\item{Y}{The numeric vector \eqn{\large \bar{Y}}, such that \eqn{\large y[k] \in \bar{Y}}}

\item{y0}{The term, \eqn{\large y_0}, acting as the constant of integration}
}
\value{
The returned value is always a two-column dataframe with columns \code{X} and \code{Y}
\code{X} contains the values \eqn{x \in \bar{X}}
\code{Y} contains the values \eqn{y_1 \in \bar{Y}}
}
\description{
The function takes in equi-dimensional vectors, \eqn{\large \bar{X}} and \eqn{\large \bar{Y}} and solves the equation shown below
\deqn{\displaystyle \Large y_1[k] = \int_{x_0}^{x[k]}f(x)\,\mathrm{d}x + y_0 \quad | \quad y = f(x)}\cr
It then returns the dataframe consisting of vectors \eqn{\large \bar{X}} and \eqn{\large \bar{Y_1}}
}
\details{
The function uses the Trapezoidal rule for calculating the integral. Please see \link{integrate.num}
}
\examples{
rm(list=ls())
library(ImageMetrologyAnalysis)
#Build the dataset
X <- seq(-4*pi, 4*pi, by= 0.001)
Y <- cos(X)
#Plot the function
plot(X, Y, 'l')
#Call the integrate function
df <- integrate.function(X, Y, y0=0)
#Plot the data frame
lines(df$X, df$Y, col = 'red')
#Use vertical lines as indicators of the phase shift between sine and cosine
abline(v=c(0, pi/2, -pi/2))


}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
