% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.num.R
\name{integrate.num}
\alias{integrate.num}
\title{Integrates a 1D dataset between specific bounds}
\usage{
integrate.num(X, Y, xmin, xmax, riemann)
}
\arguments{
\item{X}{The numeric vector which contains the \eqn{\large x} values or the independent variable}

\item{Y}{The numeric vector which contains the \eqn{\large y = f(x)} values or the dependent variable}

\item{xmin}{The lower bound of the integral, denoted as \eqn{\large x_0}}

\item{xmax}{The upper bound of the integral, denoted as \eqn{\large x_1}}

\item{riemann}{A logical bit which toggles between a riemann summation or a trapezoidal approximation.\cr
Selects a riemann integral when set (\code{TRUE}).}
}
\value{
The returned value is always a numeric scalar
}
\description{
The function provides the closest analogue of the deinite integral shown below
\deqn{\large\displaystyle S = \int_{x_0}^{x_1}f(x)\,\mathrm{d}x}
}
\details{
For additional details between a Trapezoidal approximation and a Riemann summation, see also \link[StatsChitran]{num_integrate}
, \link[StatsChitran]{num.integrate.riemann}.\cr
For very dense vectors, \eqn{\large \bar{X}} and \eqn{\large \bar{Y}},  a Riemann integral approaches a Trapezoidal in precision estimates.
}
\examples{
#Call the relevant libraries
library(StatsChitran)
library(ImageMetrologyAnalysis)

rm(list = ls())
X <- seq(-10, 10, by=0.005)
Y <- X**3
plot(X, Y, type = 'l')
integrate.num(X, Y, xmin = 0, xmax = 5, riemann = F)


}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
