% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.tensor.rank3.R
\name{is.tensor.rank3}
\alias{is.tensor.rank3}
\title{Plottable rank 3 tensors}
\usage{
is.tensor.rank3(arr)
}
\arguments{
\item{arr}{Rank 3 numeric array with further checks}
}
\value{
A boolean \code{TRUE} or \code{FALSE} bit
}
\description{
checks whether the object is a rank 3 tensor that can be plotted
}
\details{
Checks whether the tensor is of type \eqn{T_{m, n, p}} used by \link{StatsChitran::plot2D.arr}
The following checks are run by the program
\itemize{
\item If object is of type array
\item If \code{dim(arr)} returns a vector of length 3
\item If the 3rd index of the array \eqn{p} is equal to 3
\item If the Tensor slices \eqn{T_{m,n,1}, \,\, T_{m, n, 2} \dots} are individually numeric using \code{is.numeric()}
\item If the X and Y slices \eqn{T_{m, n, 2}, \, \, T_{m, n, 3}} are rank 1 matrixes
\item It is important to note that these conditions are necessary but not sufficient for it to be a plottable array
}
}
\examples{
library(StatsChitran)
dat <- system.file("extdata", "STM.png", package = "StatsChitran")
tens <- img_2_arr(source.png = dat, x.lim = c(0, 6.115*10^-9), y.lim = c(0, 6.115*10^-9))
is.tensor.rank3(tens)


}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
