% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-latt_reconstr_1axis.R
\docType{data}
\name{latt_reconstr_1axis}
\alias{latt_reconstr_1axis}
\title{A lattice showing reconstruction along one dimension}
\format{
A numeric array with dimensions (401, 401, 3)
}
\source{
Generated using internal R scripts
}
\usage{
latt_reconstr_1axis
}
\description{
A numeric 3D array of dimensions (401 × 401 × 3), generated internally
and used for example testing of fft_2D_map() function.
}
\keyword{datasets}
