% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin.bg.sub.1D.R
\name{lin.bg.sub.1D}
\alias{lin.bg.sub.1D}
\title{Linear background subtraction}
\usage{
lin.bg.sub.1D(dat, win, min.zero)
lin.bg.sub.1D(dat, win)
}
\arguments{
\item{dat}{The dataset from which the background is to be subtracted. \cr
The format for the entry is shown below
\itemize{
\item Must be a two column numeric dataframe
\item First column must be the X values and the 2nd column the Y values
}}

\item{win}{The windows/subsets w.r.t X over which the background is to be calculated for subtraction. \cr
The format for the entry is shown below
\itemize{
\item Must be a two column numeric dataframe
\item Each row consists of the minimum(1st column) and maximum(2nd column) X-values of a two column dataframe
\item Each minimum value should be less than each max value in the same row. Else an error is thrown.
\item Please see \link{Examples}
}}

\item{min.zero}{A boolean bit which when set to \code{TRUE} ensures that the minimum value of the returned dataframe's Y values will be clamped at zero.\cr
Defaults to \code{TRUE}}
}
\value{
A Two column dataframe similar in shape, size, topology to the \code{dat} dataframe
}
\description{
Subtracts a linear background from line profiles.\cr
Based on the input dataset from which the background is to be subtracted and...\cr
the windows over which the background is to be calculated
}
\examples{
#Call the relevant libraries
library(StatsChitran)
library(ImageMetrologyAnalysis)

#build the dataset
X <- seq(-10, 10, 0.1)
Y <- gauss(X, sig = 2, probability = T)
Y <- Y + rnorm(length(X), sd=0.005)

#add bg
Y <- Y + 0.02*X
plot(X, Y)
#figure out the windows over which to calculate the background
##window 1
abline(v=c(-10, -5), col='red')
##window 2
abline(v=c(5, 10), col='red')
## build the win dataframe
win <- matrix(c(-10, -5, 5, 10), byrow = T, nrow = 2)
win <- as.data.frame(win)

#Call the lin.bg.sub.1D() function
bg.sub.dat <- lin.bg.sub.1D(dat = data.frame(X, Y), win = win, min.zero = T )
plot(bg.sub.dat$X, bg.sub.dat$Y)


}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
