% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin.prof.h.R
\name{lin.prof.h}
\alias{lin.prof.h}
\title{Extract horizontal line profiles from \eqn{T_{m,n,p=3}} type tensors}
\usage{
lin.prof.h(tens, h.val, F)
lin.prof.h(tens, h.val)
}
\arguments{
\item{tens}{The type \eqn{\large T_{m,n,p=3}} tensor from which the line profile is expected \cr}

\item{h.val}{The value at which the expected line profile is to be extracted}

\item{pl}{A boolean bit which when set to \code{TRUE} ensures that the tensor is plotted using \link{StatsChitran::plt.2D.arr}\cr
Defaults to \code{TRUE}}
}
\value{
A Two column dataframe whose no. of rows are equal to \eqn{\large n}
\itemize{
\item The first column are the distance values
\item The second column are the height/magnitude values
}
}
\description{
Extracts horizontal line profiles from a Tensor of the type \eqn{\large T_{m,n,p=3}}\cr
For more information about \eqn{\large T_{m,n,p=3}} type tensors, look up \link{StatsChitran::img_2_arr}
}
\examples{
rm(list=ls())
library(StatsChitran)
library(ImageMetrologyAnalysis)
X <- seq(-10, 10, by = 0.05)
Y <- seq(-10, 10, by=0.05)
# Create two tensors using gaussians
G_tens <- gauss.2D.nocorr(X, Y, mu.x = 3.5, mu.y = 2.5, sig.x = 1.7, sig.y = 2.3) #1st tensors
G1_tens <- gauss.2D.nocorr(X,Y) #2nd tensor
G_tens[,,1] <- G_tens[,,1]+G1_tens[,,1] #Add the two tensors
plot2D.arr(G_tens) #Plot the tensor

#Call the lin.prof.h() function
df <- lin.prof.h(G_tens, h.val = 3.5)

#Plot the extracted line profile
plot(df$dist, df$magnitude, type = 'l')


}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
