% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalizes/Recalibrates a matrix between 0 and 1}
\usage{
normalize(M)
}
\arguments{
\item{M}{Numeric matrix that you want to normalize.}
}
\value{
A numeric matrix of the same dimension as \code{M}.

#' @author
Chitran Ghosal <ghosal.chitran@gmail.com>
}
\description{
This function takes in a matrix and outputs a linearly recalibrated matrix where the minimum value is recalibrated to 0 and the max value is reclaibrated to 1
}
\examples{
M = matrix(seq(1,10), nrow=2)
normalize(M)
}
