% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.boxes.R
\name{plot2D.boxes}
\alias{plot2D.boxes}
\title{Plot 2D boxes in 2-dimensions}
\usage{
plot2D.boxes(img.tens, box.mat, box.thick, box_intens, pl)
plot2D.boxes(img.tens, box.mat, box.thick, box_intens)
}
\arguments{
\item{img.tens}{The tensor describing the lattice over which the boxes need to be drawn}

\item{box.mat}{The matrix defining the co-ordinaes of the boxes that have to be drawn
\itemize{
\item The no. of rows = no. of boxes needed
\item The no. of columns should be equal to 4
\item Each row should be of the form \code{c(x_LL, y_LL, x_UR, y_UR)}
\item Here \code{x_LL, y_LL} signify the lower left corner x and y co-ordinate of the rectangular box respectively
\item Here \code{x_UR, y_UR} signify the upper right corner x and y co-ordinates of the rectangular box respectively
}}

\item{box.thick}{The thickness of the box w.r.t the scale of the x and y co-ordinate respectively}

\item{box_intens}{The intensity factor which decides the color of the box in grayscale
\itemize{
\item An intensity factor of \code{1.0} implies a grayscale value equal to the highest value residing in \code{img.tens}
}}

\item{pl}{Plots \code{img.tens} with the boxes drawn on them if set to \code{TRUE}.\cr
Defaults to \code{TRUE}}
}
\value{
The returned value is always a rank-3 tensor.\cr
\itemize{
\item The X and Y slices of the tensor \eqn{\large T_{m, n, p=2}} and \eqn{\large T_{m,n, p=3}} respectively stay the same
}
}
\description{
The function takes in a rank-3 tensor of the form \eqn{\large T_{m,n,p = 3}} and returns the tensor with boxes drawn around specific co-ordinates
}
\examples{
#call the relevant libraries
library(StatsChitran)
library(ImageMetrologyAnalysis)

#draw the lattice in question
X <- seq(-10, 10, by= 0.05)
Y <- X
latt_sq <- sq.latt(X, Y, R.latt = 1.0, A=1, sig = 0.2)

#build the fourier transform of the lattice in question
f.latt_sq <- fft_2D(latt_sq, sampling.del = 0.1, pl='amp')
#select the amplitude spectrum
f.latt_amp <- f.latt_sq[[1]]
#zoom into the area of interest
f.latt_amp_interest <- plot2D.zoom(f.latt_amp, center = c(0,0), Del_X = 20, Del_Y = 20)


###Estimate the co-ordinates for the points
plot2D.arr(f.latt_amp_interest)
df <- lin.prof.h(f.latt_amp_interest, h.val = -0.25)
df <- lin.prof.v(f.latt_amp_interest, v.val = 6.5)

###Build the matrix box.mat as b.mat and then check it
b.vec <- c(5.0, -1.5, 8.0, 1.5, -1.5, 4.5, 1.5, 7.5 )
b.mat <- matrix(data = b.vec, byrow = T, ncol = 4)
###Plot the boxes
dat <- plot2D.boxes(f.latt_amp_interest, b.mat, box.thick = 0.025, box_intens = 0.2)


}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
