% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.win.sig.R
\name{plot2D.win.sig}
\alias{plot2D.win.sig}
\title{Sigmoid based windowing of an image Tensor}
\usage{
plot2D.win.sig(tens, center, k, Xspan, Yspan, pl)
plot2D.win.sig(tens, center, k, Xspan, Yspan)
}
\arguments{
\item{tens}{The image tensor of the form \eqn{\large T_{m,n,p=3}} that you want to window a part of}

\item{center}{The co-ordinates of the center of the window
\itemize{
\item Must be a 2-element numeric vector of the form \code{center = c(Xc, Yc)}
\item Here \code{Xc} and \code{Yc} are the X and Y co-ordinates of the center
}}

\item{k}{smoothness/sharpness parameter of the sigmoid function being used for the windowing
\itemize{
\item Higher is sharper and lower is smoother/softer
\item Please see \link{sigmoid.2D.plat}, \link{sigmoid.plat}
}}

\item{Xspan, Yspan}{The spans around the center of the window. Numeric scalars.}

\item{pl}{Boolean bit. Permission to plot the resulting window
Defaults to \code{TRUE}}
}
\value{
The returned value is always a rank-3 image tensor \eqn{\large T_{m,n,p=3}}\cr
It has the exact same dimensions as the \code{tens} argument \cr
It has the slice-2 and slice-3 matrixes untouched since they are the X-Y grid of the image
}
\description{
Window a certain area of a rank-3 image tensor using the sigmoid function.
}
\details{
Please see \link{sigmoid}, \link{sigmoid.plat}, \link{sigmoid.2D.plat} for a better understanding\cr
Useful for windowing areas of interest in lattices for fourier/spectral analysis
}
\examples{
#Call the relevant libraries
library(StatsChitran)
library(ImageMetrologyAnalysis)

#Load the lattice
data("latt_reconstr_1axis", package = "ImageMetrologyAnalysis")
latt <- latt_reconstr_1axis
plot2D.arr(latt)
latt.zoom <- plot2D.win.sig(tens = latt, center = c(5,0), k = 10, Xspan = 8, Yspan = 2)


}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
