% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rect.latt.R
\name{rect.latt}
\alias{rect.latt}
\title{Build a rank-3 tensor pertaining to a periodic rectagular lattice}
\usage{
rect.latt(X, Y, R.latt.x, R.latt.y, A, sig, pl)
rect.latt(X, Y, R.latt.x, R.latt.y, A, sig)
}
\arguments{
\item{X, Y}{The \eqn{\large \bar{X}} and \eqn{\large \bar{Y}} vectors pertaining to the image spans\cr
See \code{\link{grid_2}} for usage specifics of the X and Y vectors}

\item{R.latt.x, R.latt.y}{The lattice constants \eqn{\large \bar{R}_1} and \eqn{\large \bar{R}_2}}

\item{A, sig}{The amplitude and standard deviation of the gaussian to be placed at each lattice site}

\item{pl}{When set to \code{TRUE}, plots the rank-3 tensor using \code{\link{StatsChitran::plot2D.arr}}\cr
Defaults to \code{TRUE}}
}
\value{
The returned value is always a rank-3 tensor.\cr
\itemize{
\item The first slice, \eqn{\large T_{m,n, p = 1}}, are the lattice Z-values
\item The second slice, \eqn{\large T_{m,n, p = 2}}, contains the X-values of the X-Y grid
\item The third column, \eqn{\large T_{m,n, p = 3}}, contains the Y-values of the X-Y grid
}
}
\description{
The function takes in vectors \eqn{\bar{X}}, \eqn{\bar{Y}} and lattice costants \eqn{\large \bar{R}_1}, \eqn{\large \bar{R}_2} and returns the rank-3 tensor pertaining to the lattice.\cr
}
\details{
Creates a periodic lattice by placing gaussians at each lattice site within the \eqn{\large \bar{X}}, \eqn{\large \bar{Y}} span\cr
The lattice periodicity is obviously defined by \eqn{\large \bar{R}_1} and \eqn{\large \bar{R}_2}\cr
}
\examples{
rm(list = ls())
library(ImageMetrologyAnalysis)
library(StatsChitran)
X <- seq(-5, 5, by=0.01)
Y <- seq(-3, 3, by = 0.01)
rec.tens <- rect.latt(X, Y, R.latt.x = 0.5, R.latt.y = 0.75, A = 1, sig = 0.15)


}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
