% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy_clean.R
\name{roxy_clean}
\alias{roxy_clean}
\title{Cleans the manual files for roxygen2}
\usage{
roxy_clean()
}
\value{
The function returns nothing
}
\description{
The function deletes the manual files for the package and then recreates them using \code{roxygen2::roxygenize()}
}
\examples{
\dontrun{
roxy_clean()
}



}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
