% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.2D.plat.R
\name{sigmoid.2D.plat}
\alias{sigmoid.2D.plat}
\title{The sigmoid.2D.plat function}
\usage{
sigmoid.2D.plat(X, Y, k, x.left, x.right, y.left, y.right, pl)
sigmoid.2D.plat(X, Y, k, x.left, x.right, y.left, y.right)
}
\arguments{
\item{X, Y}{The numeric vectors which create the XY grid}

\item{k}{The sharpness \eqn{k} of the sigmoid function
\itemize{
\item higher \eqn{k} implies a sharper cutoff
}}

\item{x.left, x.right, y.left, y.right}{The cutoff values along the X and Y directions}

\item{pl}{Plots the 2D sigmoid plateau when set to \code{TRUE}\cr
Defaults to \code{TRUE}}
}
\value{
The returned value is always a numeric matrix
}
\description{
The sigmoid plateau function in two dimensions
}
\details{
The function returns a numeric matrix \code{Z(n,m)}\cr
\code{n = size(Y)}\cr
\code{m = size(X)}\cr
The matrix is built as a product of sigmoid plateaus running along the orthogonal X and Y directions\cr
}
\examples{
#Call the relevant libraries
library(ImageMetrologyAnalysis)
library(StatsChitran)
#Create the sequences
X <- seq(-10, 10, by= 0.01)
Y <- seq(-7, 7, by=0.01)
#Create the plateau matrix and plot it
Z <- sigmoid.2D.plat(X, Y, k=3, x.left = -8, x.right = 8, y.left = -3, y.right = 3)


}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
