% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.plat.R
\name{sigmoid.plat}
\alias{sigmoid.plat}
\title{The simplified sigmoid function}
\usage{
sigmoid.plat(X, k, left.cut, right.cut)
}
\arguments{
\item{X}{The numeric vector which contains the \eqn{x} values against which the sigmoid vector is needed}

\item{k}{The sharpness \eqn{k} of the sigmoid function
\itemize{
\item higher \eqn{k} implies a sharper cutoff
}}

\item{left.cut}{The cutoff value \eqn{x_1} at the left side of the sigmoid plateau}

\item{right.cut}{The cutoff value \eqn{x_2} at the right side of the sigmoid plateau}
}
\value{
The returned value is always a vector
}
\description{
The sigmoid plateau function in one dimension
}
\details{
The function returns a vector \eqn{\large \bar{Y}} with individual components \eqn{\large y} against the vector \eqn{\large \bar{X}} with component \eqn{\large x}
\deqn{\Large y = \{1/(1 + e^{-k(x - x_1)})\} - \{1/(1 + e^{-k(x - x_2)})\} }
}
\examples{
#Call the relevant libraries
library(ImageMetrologyAnalysis)

#Build the X-variable and then call the sigmoid.plat() function
X <- seq(-10, 10, by=0.01)
Y1 <- sigmoid.plat(X, k = 7, left.cut = -8, right.cut = 8)
Y2 <- sigmoid.plat(X, k = 2, left.cut = -8, right.cut = 8)

#Plot the curves
plot(X, Y1, type = 'l', col='blue')
lines(X, Y2, type = 'l', col='green')
abline(v=c(-8,8), col='red')


}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
