% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{The simplified sigmoid function}
\usage{
sigmoid <- function(X, k, cutoff)
}
\arguments{
\item{X}{The numeric vector which contains the \eqn{x} values against which the sigmoid vector is needed}

\item{k}{The sharpness \eqn{k} of the sigmoid function
\itemize{
\item higher \eqn{k} implies a sharper cutoff
}}

\item{cutoff}{The cutoff value \eqn{x_0} at which the sigmoid function returns a value of \eqn{1/2}}
}
\value{
The returned value is always a vector
}
\description{
The sigmoid function in one dimension
}
\details{
The function returns a vector \eqn{\large \bar{Y}} with individual components \eqn{\large y} against the vector \eqn{\large \bar{X}} with component \eqn{\large x}
\deqn{\Large y = 1/(1 + e^{-k(x - x_0)})}
}
\examples{
#Import the relevant libraries
library(StatsChitran)
library(ImageMetrologyAnalysis)

#Build the X-variable
X <- seq(-10, 10, by=0.01)
#Build the Y-variable
Y1 <- sigmoid(X, k=1, cutoff = 3)
Y2 <- sigmoid(X, k=5, cutoff = 3)
#Plot the results
plot(X, Y1, type='l', col='blue')
lines(X, Y2, type = 'l', col='green')
abline(v=3, col='red')


}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
