% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_sym_auto.R
\name{solve_sym_auto}
\alias{solve_sym_auto}
\title{Solves a system of Linear Equations}
\usage{
solve_sym_auto(A, b)
}
\arguments{
\item{A}{Numeric matrix for the system of linear equations}

\item{b}{Numeric Vector having length equal to no. of equations}
}
\value{
A numeric vector of the same length as \eqn{[b]}.
}
\description{
This function takes in a non-singular square matrix and a vector of the same length as the no. of rows of the matrix \\cr
The no. of conditions/equations are equal to the no. of rows of matrix \code{A}
}
\details{
Solves the matrix equation as shown below. \cr
\deqn{ [A] [X] = [b]; \quad \implies \text{returns} \,\, [X] = [A]^{-1}[b]}
For example, the equations shown below
\deqn{
x + y = 3 \\\\
x - y = 1
}
Can be arranged as shown
\deqn{
\begin{pmatrix}
1 & 1 \\
1 & -1
\end{pmatrix}\cdot
\begin{pmatrix}
x \\
y
\end{pmatrix} =
\begin{pmatrix}
3\\
1
\end{pmatrix}
}
Where \eqn{[A]} and \eqn{[b]} are represented as shown below
\deqn{
[A] =
\begin{pmatrix}
1 & 1\\
1 & -1
\end{pmatrix}
\quad [b] =
\begin{pmatrix}
3\\
1
\end{pmatrix}
}
}
\examples{
A <- matrix(data = c(1,1,1, -1), nrow = 2, byrow = T)
b <- c(3,1)
solve_sym_auto(A,b)

}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
