% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq.latt.R
\name{sq.latt}
\alias{sq.latt}
\title{Build a rank-3 tensor pertaining to a periodic square lattice}
\usage{
sq.latt(X, Y, R.latt, A, sig, pl)
sq.latt(X, Y, R.latt, A, sig)
}
\arguments{
\item{X, Y}{The \eqn{\large \bar{X}} and \eqn{\large \bar{Y}} vectors pertaining to the image spans\cr
See \code{\link{grid_2}} for usage specifics of the X and Y vectors}

\item{R.latt}{The lattice constant \eqn{\large \bar{R}} for the square lattice}

\item{A, sig}{The amplitude and standard deviation of the gaussian to be placed at each lattice site}

\item{pl}{When set to \code{TRUE}, plots the rank-3 tensor using \code{\link{StatsChitran::plot2D.arr}}\cr
Defaults to \code{TRUE}}
}
\value{
The returned value is always a rank-3 tensor.\cr
\itemize{
\item The first slice, \eqn{\large T_{m,n, p = 1}}, are the lattice Z-values
\item The second slice, \eqn{\large T_{m,n, p = 2}}, contains the X-values of the X-Y grid
\item The third column, \eqn{\large T_{m,n, p = 3}}, contains the Y-values of the X-Y grid
}
}
\description{
The function takes in vectors \eqn{\bar{X}}, \eqn{\bar{Y}} and lattice costant \eqn{\large \bar{R}} and returns the rank-3 tensor pertaining to the lattice.\cr
}
\details{
Creates a periodic lattice by placing gaussians at each lattice site within the \eqn{\large \bar{X}}, \eqn{\large \bar{Y}} span\cr
The lattice periodicity is obviously defined by \eqn{\large \bar{R}}\cr
}
\examples{
rm(list = ls())
library(ImageMetrologyAnalysis)
library(StatsChitran)
X <- seq(-10, 10, by=0.05)
Y <- seq(-5, 5, by = 0.05)
sq.tens <- sq.latt(X = X, Y = Y, R.latt = 0.75, A = 1, sig = 0.1)


}
\author{
Chitran Ghosal \href{mailto:ghosal.chitran@gmail.com}{ghosal.chitran@gmail.com}
}
